Use common logic for rhs of assignment expressions
