+++ Src/CSharpier/SyntaxPrinter/SyntaxNodePrinters/VariableDeclaration.cs
@@ -1,6 +1,4 @@
-using System;
 using CSharpier.DocTypes;
-using CSharpier.SyntaxPrinter;
 using Microsoft.CodeAnalysis.CSharp.Syntax;
 
 namespace CSharpier.SyntaxPrinter.SyntaxNodePrinters
@@ -25,28 +23,6 @@ namespace CSharpier.SyntaxPrinter.SyntaxNodePrinters
             var variable = node.Variables[0];
             var initializer = variable.Initializer;
 
-            var formatMode = initializer?.Value
-                is AnonymousObjectCreationExpressionSyntax
-                    or AnonymousMethodExpressionSyntax
-                    or InitializerExpressionSyntax
-                    or InvocationExpressionSyntax
-                    or ConditionalExpressionSyntax
-                    or ObjectCreationExpressionSyntax
-                    or SwitchExpressionSyntax
-                    or LambdaExpressionSyntax
-                ? FormatMode.NoIndent
-                : initializer?.Value
-                        is BinaryExpressionSyntax
-                            or InterpolatedStringExpressionSyntax
-                            or IsPatternExpressionSyntax
-                            or LiteralExpressionSyntax
-                            or QueryExpressionSyntax
-                            or StackAllocArrayCreationExpressionSyntax
-                        ? FormatMode.IndentWithLine
-                        : FormatMode.Indent;
-
-            var groupId = Guid.NewGuid().ToString();
-
             return Doc.Concat(
                 Doc.Group(
                     Node.Print(node.Type),
@@ -59,46 +35,8 @@ namespace CSharpier.SyntaxPrinter.SyntaxNodePrinters
                         ? Doc.Concat(" ", Token.Print(initializer.EqualsToken))
                         : Doc.Null
                 ),
-                initializer != null
-                    ? Doc.Concat(
-                            formatMode is FormatMode.IndentWithLine
-                                ? Doc.Null
-                                : Doc.GroupWithId(groupId, Doc.Indent(Doc.Line)),
-                            initializer.Value
-                                is InvocationExpressionSyntax
-                                    or ParenthesizedLambdaExpressionSyntax
-                                    or ObjectCreationExpressionSyntax
-                                    or ElementAccessExpressionSyntax
-                                    or ArrayCreationExpressionSyntax
-                                    or ImplicitArrayCreationExpressionSyntax
-                                ? Doc.IndentIfBreak(
-                                        Doc.Group(Node.Print(initializer.Value)),
-                                        groupId
-                                    )
-                                : Doc.Group(IndentIfNeeded(initializer.Value, formatMode))
-                        )
-                    : Doc.Null
+                initializer != null ? RightHandSide.Print(initializer.Value) : Doc.Null
             );
-        }
-
-        private static Doc IndentIfNeeded(ExpressionSyntax initializerValue, FormatMode formatMode)
-        {
-            if (formatMode is FormatMode.NoIndent)
-            {
-                return Node.Print(initializerValue);
-            }
-
-            return Doc.Indent(
-                formatMode is FormatMode.IndentWithLine ? Doc.Line : Doc.Null,
-                Node.Print(initializerValue)
-            );
-        }
-
-        private enum FormatMode
-        {
-            NoIndent,
-            IndentWithLine,
-            Indent
         }
     }
 }
